update: isolate types to files
